Refactoring response code and entity
