+++ src/main/java/ru/yandex/practicum/filmorate/controller/ErrorResponse.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.filmorate.controller;
+
+import lombok.Value;
+
+@Value
+public class ErrorResponse {
+    String error;
+}
