+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -2,8 +2,8 @@ package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.validation.ValidationException;
 
 import javax.validation.Valid;
 import java.util.ArrayList;
@@ -30,11 +30,12 @@ public class UserController {
 
     @PostMapping
     public User add(@Valid @RequestBody User user) {
-        log.info("Adding user: {}", user);
-
         User newUser = user.toBuilder()
                 .id(nextId.incrementAndGet())
                 .build();
+
+        log.info("Adding user: {}", newUser);
+
         map.put(newUser.getId(), newUser);
         return map.get(newUser.getId());
     }
@@ -47,7 +48,7 @@ public class UserController {
             map.put(user.getId(), user);
             return map.get(user.getId());
         } else {
-            throw new ValidationException("id", String.format("User with id=%d not found", user.getId()));
+            throw new NotFoundException(String.format("User with id=%d not found", user.getId()));
         }
     }
 }
