+++ src/main/java/ru/yandex/practicum/filmorate/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.filmorate.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
