+++ src/main/java/ru/yandex/practicum/filmorate/exception/ValidationException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.validation;
+package ru.yandex.practicum.filmorate.exception;
 
 import lombok.Getter;
