+++ src/main/java/ru/yandex/practicum/filmorate/validation/Violation.java
@@ -2,9 +2,7 @@ package ru.yandex.practicum.filmorate.validation;
 
 import lombok.Getter;
 import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
 @Getter
 @ToString
 public class Violation {
@@ -14,6 +12,5 @@ public class Violation {
     public Violation(String fieldName, String message) {
         this.fieldName = fieldName;
         this.message = message;
-        log.warn(this.toString());
     }
 }
