+++ src/test/java/ru/yandex/practicum/filmorate/controller/UserControllerTest.java
@@ -86,8 +86,7 @@ class UserControllerTest {
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(mapper.writeValueAsString(newUser)))
                 .andDo(print())
-                .andExpect(status().is4xxClientError())
-                .andExpect(jsonPath("$.violations[0].fieldName", Matchers.equalTo("id")));
+                .andExpect(status().is4xxClientError());
     }
 
     @Test
